+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version: 3.2.0
 ENHANCEMENTS:
 
 * Improvement done to determine the step line number associated with each image in gallery.
+* Improvement done in the way hooks are executed, to properly handle the case where BehaveX hooks are not initialized yet.
 
 Version: 3.1.1
 -------------------------------------------------------------------------------